backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
SQLAlchemy>=2.0,<3
PyMySQL>=1.1,<2
cryptography>=43
bcrypt>=4.0,<5
PyJWT>=2.8,<3
pydantic-settings>=2.6,<3
python-dotenv>=1.0,<2
python-multipart>=0.0.9,<1
